implemented file resize transformation + refactor on fetch file from s3

+++ src/main/java/org/gustavolyra/image_process_service/models/dto/transformations/TransformationsDto.java
@@ -5,10 +5,9 @@ import lombok.Data;
 @Data
 public class TransformationsDto {
 
-    private Double rotate;
+    private int rotate;
     private String format;
     private Crop crop;
     private Resize resize;
     private Filters filters;
-
 }
